+++ src/Exceptions/ImageIsAlreadyCachedException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Kwaadpepper\ImageResizer\Exceptions;
+
+/**
+ * Used for internal work, happens when a resize imageg is in cache.
+ */
+class ImageIsAlreadyCachedException extends \ErrorException
+{
+}