+++ src/Exceptions/ImageResizerException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Kwaadpepper\ImageResizer\Exceptions;
+
+/**
+ * Will occur if file is not readable or format is invalid.
+ */
+class ImageResizerException extends \ErrorException
+{
+}