+++ src/helpers.php
@@ -7,11 +7,11 @@ use Kwaadpepper\ImageResizer\ImageResizer;
  *
  * @param string $path
  * @param string $configName
- * @return string
+ * @return string Returns the resized file path from cache of the original file path if resize could not be done.
  */
 function resize(string $path, string $configName = null): string
 {
-    return ImageResizer::resizeImageOrIgnore($path, $configName);
+    return ImageResizer::resizeImageOrIgnore($path, $configName) ?? $path;
 }
 
 /**
@@ -19,9 +19,9 @@ function resize(string $path, string $configName = null): string
  *
  * @param string $path
  * @param string $configName
- * @return string
+ * @return string Returns the resized file path from cache of the original file path if resize could not be done.
  */
 function resizePublic(string $path, string $configName = null): string
 {
-    return resize(public_path($path), $configName);
+    return ImageResizer::resizeImageOrIgnore(public_path($path), $configName) ?? $path;
 }